dodawanie dokumentacja

+++ src/lib/Coord.hpp
@@ -12,7 +12,10 @@
 
 using json = nlohmann::json;
 
-typedef int64_t CoordDim;
+/**
+ * @brief Typ liczby całkowitej używanej do obliczeń
+ */
+typedef int16_t CoordDim;
 
 /**
  * @brief Zamienia wymiar na tekst przyjazny systemowi plików (1 -> p1, -1 -> m1)
